Remove storage permission and bump to 1.0.1.
For now, this permission is unused, so don't declare it. It's not listed
as sensitive per:

https://developer.chrome.com/docs/extensions/mv3/permission_warnings/#permissions_with_warnings

so it's unlikely to disable the extension if added later, though we
could be safe and add it as optional and prompt if we need to save some
state (though this will close the popup and require it to be reopened).