Introduce a new source for scraping .puz links.
This is a minor change meant to make it more convenient to download
.jpz/.pdf files when only .puz files are provided. However, it comes
with a few broader changes:

- Since .puz links can be hosted on any domain, we now specify that we
can request any host permission dynamically, and rework the permission
requesting scheme to allow a source to inspect the current tab contents
before determining which host permissions are necessary.

- Since the same page might have both .puz links and an applet, we do
some rough deduplication by comparing the grids of scraped puzzles.