Defer file format conversion until a link is clicked.
This allows us to show the list of available puzzles far more quickly,
and avoids unnecessary expensive conversions unless the user actually
elects to download a file.

One drawback of this approach is that you can no longer right click a
link and click "Save As".

Another drawback is that we might show an impossible conversion, e.g. a
puzzle that cannot be represented as a PUZ or PDF file due to use of
some special feature. However, it's probably better to show the link and
show an error if conversion fails than to hide the link without visible
explanation.