Add source for Crossword Compiler applets.
Supports any applet embedded directly on the current page, as well as
sites which use an iframe to embed a page with the applet as long as the
page comes from a known list. (Otherwise, we'd need to request global
host permissions in order to be able to check for the presence of an
applet on an iframe).

See #6