Add support for Manifest V3.
Build targets are changed to from development and production to
developmentV{2,3} and productionV{2,3}. While the Javascript code is the
same - we check the manifest version at runtime and run different logic
as needed - we need a different manifest.json file to declare the new
version and update to match the changes in V3. For now, we use
placeholders in the manifest with different values for V2 and V3. In the
future, we might merge these flavors, but introduce separate ones for
Firefox and Chrome which will need to declare different minimum
versions.

The primary functional change is that the API to execute scripts on the
current tab has moved and now requires that the code to run be a
predefined script or function rather than an arbitrary code snippet.
References to Kotlin methods don't work, but we can use js() to define
functions and execute those. For backwards compatibility, we can
serialize those functions to strings and inject them to update a
temporary DOM object that we read back, as before.

This will not be pushed to Chrome until closer to the deadline, since it
depends on the recent version 102 for optional_host_permissions, and we
should give users more time to update.

Outstanding issues:

-  We use chrome.scripting.executeScript directly instead of a browser
   polyfill since none is currently available.
-  The font files are exposed to all sites instead of just the extension
   itself. This matches MV2 behavior, but could be locked down further.
   However, it's unclear how to reference our own extension ID when it
   is dynamic.
-  We add "(Manifest v3)" to the version in the pop-up so it can be
   distinguished from the v2 extension when both are being tested.

See #4