Remove obsolete webextension-polyfill hacks.
browser.scripting already exists and works via webextension-polyfill;
the calls are just proxied directly through to the Chrome
implementation since it already returns a Future. All that is missing
upstream are integration tests.

We also no longer need the workaround to catch non-Errors passed to
Promise.resolve, since this was an issue in webextension-polyfill that
was fixed there.

We still need to define browser.scripting if not already defined to
prevent crashing on Manifest v2.