Add option to automatically download puzzles.
When enabled, and there is exactly one puzzle on a page (that is
successfully scraped), automatically download the puzzle in the user's
selected format and close the popup. Show the popup as normal if there
are zero puzzles, more than one puzzle, or if there was an error
downloading in the selected format (in which case a dialog is shown on
top of the list of formats).

This mostly works as naturally implemented. It was not feasible to
implement a context menu option which force-opens the popup because
there is no way to open the popup from a background page - the user must
click on it. There is also a hack to wait 100ms after initiating the
download which seems to be necessary for Firefox to reliably download
the file - no such delay is necessary in Chrome, and shorter delays in
Firefox appear unreliable. Ideally, we could wait for the download to
start; the only potential way to make this work would be to use the
browser.downloads API to perform the download and listen for it to fully
complete (the start event does not appear to suffice, from testing).

See #21