Fix Puzzle Society scraping.
The puzzle data no longer appears to be embedded in a global variable on
the page. We take an approach similar to Crosshare - look in the
embedded window.__NEXT_DATA__ variable for a URL that contains the raw
puzzle data, and if it points to the wrong puzzle (due to navigating to
another puzzle from the first one), then refetch the page at the current
URL and check __NEXT_DATA__ there instead.