Remove references to browser.tabs.executeScript from Chrome build.
The extension is currently being rejected because "an external file is
injected in the js/CrosswordScraper.js file". While this isn't true,
my best guess is that the (unused) call to executeScript, only needed
for Firefox, is tripping up the checks. Remove it by factoring out a
browser-specific function for executing our (included) scripts, so that
the only method included with Chrome is the permitted one which takes an
actual function instead of an arbitrary string.

This is temporary until we can migrate to Manifest v3 for Firefox.