Add support for PZZL crosswords.
The puzzle data does not appear to be accessible via Javascript
directly, so we download it via the same API as the applet, using the
URL to determine the date.

Each PZZL applet has slightly customized Javascript and is hosted on a
different server, so we can't quite use a general purpose mechanism. For
now, we specifically handle the two known uses of this applet. If more
instances come up in the future, we could consider fetching the page
Javascript and inspecting it for the API URL, though we might still need
a list of supported domains.

See #35