Make Firefox scraping robust against Content-Security-Policy.
Due to an ancient bug
(https://bugzilla.mozilla.org/show_bug.cgi?id=1267027), our injected
scraping script can be blocked from executing on sites that set a CSP on
Firefox. A workaround is to put the script in a temporary blob and then
reference it via blob: URI.

This has the side effect of making the script's execution asynchronous,
so we also need to poll the DOM until the script is complete. Ideally,
we'd use message passing from the script to the extension to avoid
polling, but in my testing this didn't seem to work as the message would
never be received by the extension.